requirements.txt
pillow
pytesseract
requests
jsonschema
ttkbootstrap
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv